requirements.txt
Flask==1.1.2
google-cloud-secret-manager==2.0.0
gunicorn==20.0.4
PyMySQL==0.10.1
SQLAlchemy==1.3.20

